requirements.txt
matplotlib
pandas
langchain
langchain-community
langchain-groq
uvicorn
fastapi
jinja2
python-dotenv
gradio
h11
langgraph
pyyaml
black

